+++ webrtc-sys/src/lib.rs
@@ -15,10 +15,13 @@
 #[cfg(target_os = "android")]
 pub mod android;
 pub mod apm;
+pub mod audio_mixer;
 pub mod audio_resampler;
 pub mod audio_track;
 pub mod candidate;
 pub mod data_channel;
+#[cfg(any(target_os = "macos", target_os = "windows", target_os = "linux"))]
+pub mod desktop_capturer;
 pub mod frame_cryptor;
 pub mod helper;
 pub mod jsep;