+++ livekit-protocol/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.6.0](https://github.com/livekit/rust-sdks/compare/rust-sdks/livekit-protocol@0.5.2...rust-sdks/livekit-protocol@0.6.0) - 2025-12-04
+
+### Added
+
+- *(connector)* initial service impl ([#790](https://github.com/livekit/rust-sdks/pull/790))
+
+## [0.5.2](https://github.com/livekit/rust-sdks/compare/rust-sdks/livekit-protocol@0.5.1...rust-sdks/livekit-protocol@0.5.2) - 2025-11-20
+
+### Other
+
+- Remove unused dependencies ([#761](https://github.com/livekit/rust-sdks/pull/761))
+
 ## [0.5.1](https://github.com/livekit/rust-sdks/compare/rust-sdks/livekit-protocol@0.5.0...rust-sdks/livekit-protocol@0.5.1) - 2025-10-22
 
 ### Other